+++ Courier/test/utopia/courier/test/MailReadApp.scala
@@ -4,7 +4,6 @@ import utopia.courier.controller.read.{EmailBuilder, EmailReader}
 import utopia.courier.model.Authentication
 import utopia.courier.model.read.{ImapReadSettings, PopReadSettings, ReadSettings}
 import utopia.flow.generic.casting.ValueConversions._
-import utopia.flow.generic.model.mutable.DataType
 import utopia.flow.time.TimeExtensions._
 import utopia.flow.parse.file.FileExtensions._
 import utopia.flow.util.StringExtensions._
@@ -20,8 +19,6 @@ import scala.util.{Failure, Success}
   */
 object MailReadApp extends App
 {
-	
-	
 	def ask(question: String = "") = {
 		if (question.nonEmpty)
 			println(question)