+++ Vault/src/utopia/vault/nosql/access/template/Access.scala
@@ -1,6 +1,7 @@
 package utopia.vault.nosql.access.template
 
-import utopia.vault.database.Connection
+import utopia.vault.database.{Connection, References}
+import utopia.vault.model.immutable.Table
 import utopia.vault.model.template.Joinable
 import utopia.vault.nosql.view.View
 import utopia.vault.sql.JoinType.Inner
@@ -45,4 +46,23 @@ trait Access[+A] extends View
 	         joinType: JoinType = Inner)
 	        (implicit connection: Connection) =
 		read(Some(mergeCondition(condition)), order, joins, joinType)
+	
+	/**
+	  * Finds all accessible items that are **not** linked to the specified table via a foreign key reference
+	  * from either direction.
+	  * @param table Another table
+	  * @param order Custom ordering to apply (default = None = use default ordering)
+	  * @param connection Implicit DB connection
+	  * @return Items that are not linked to the specified table
+	  */
+	def findNotLinkedTo(table: Table, order: Option[OrderBy] = None)(implicit connection: Connection) =
+		// Looks for the reference between these tables
+		References.between(this.table, table).headOption match {
+			// Case: Reference found => Converts it to a join and a search condition
+			case Some(ref) =>
+				val directionalRef = if (ref.from.table == this.table) ref else ref.reverse
+				find(directionalRef.to.column.isNull, order, Vector(directionalRef.toLeftJoin), JoinType.Left)
+			// Case: No reference found => No item is considered to be linked
+			case None => read(order = order)
+		}
 }